pyproject.toml
[tool.poetry]
name = "hyperbolic-agentkit-sample"
version = "0.1.0"
description = "A sample app that uses Hyperbolic AgentKit"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
coinbase-agentkit = "^0.1.2"
coinbase-agentkit-langchain = "^0.1.0"
langchain-openai = "0.3.1"
langgraph = "^0.2.73"
langchain-core = "^0.3.15"
pydantic = "^2.9.2"
python-dotenv = "^1.0.0"
langchain-community = "^0.3.11"
langchain-experimental = "^0.3.3"
paramiko = "^3.4.0"
duckduckgo-search = "^6.4.1"
langchain-nomic = "^0.1.4"
beautifulsoup4 = "^4.12.3"
scikit-learn = "^1.6.0"
nomic = {extras = ["local"], version = "^3.3.4"}
tiktoken = "^0.8.0"
langchain-anthropic = "^0.3.0"
chromadb = "^0.6.2"
torch = "2.2"
sentence-transformers = "^3.3.1"
google-cloud-aiplatform = "^1.38.0"
vertexai = "^0.0.1"
numpy = "<2.0.0"
gradio = "^5.12.0"
pygithub = "^2.5.0"
speechrecognition = "^3.14.1"
starlette = "^0.45.3"
uvicorn = "^0.34.0"
browser-use = "^0.1.37"
tweepy = "^4.15.0"
xgboost = "^2.0.0"
pandas = "^2.2.0"
kaggle = "^1.6.0"



[tool.poetry.group.dev.dependencies]
pytest-playwright = "^0.6.2"
chromadb = "^0.6.2"
torch = "2.2"
sentence-transformers = "^3.3.1"
google-cloud-aiplatform = "^1.38.0"
vertexai = "^0.0.1"
numpy = "<2.0.0"
gradio = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
